Src/ShopManagement/Core/ShopManagement.Core/Entities/SlideAgg/SlideEntity.cs
namespace ShopManagement.Core.Entities.SlideAgg
{
    public class SlideEntity : Base.Core.BaseEntities.BaseEntity<long>
    {
        public string Picture { get; private set; }
        public string PictureAlt { get; private set; }
        public string PictureTitle { get; private set; }
        public string Heading { get; private set; }
        public string Title { get; private set; }
        public string Text { get; private set; }
        public string BtnText { get; private set; }
     

        public SlideEntity(string picture, string pictureAlt, string pictureTitle, string heading,
            string title, string text, string btnText)
        {
            Picture = picture;
            PictureAlt = pictureAlt;
            PictureTitle = pictureTitle;
            Heading = heading;
            Title = title;
            Text = text;
            BtnText = btnText;
            
        }

        public void Edit(string picture, string pictureAlt, string pictureTitle, string heading,
            string title, string text, string btnText)
        {
            Picture = picture;
            PictureAlt = pictureAlt;
            PictureTitle = pictureTitle;
            Heading = heading;
            Title = title;
            Text = text;
            BtnText = btnText;
        }

      

      
    }
}
